# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EffiPlat 全系统 API 文档
  version: "1.0"
  description: |
    EffiPlat 平台后端 API，涵盖用户、角色、权限、职责、环境、资产、服务、业务、Bug、认证等核心模块。
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [auth]
      summary: 用户登出
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: 获取当前用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'

  # 用户管理
  /users:
    get:
      tags: [users]
      summary: 获取用户列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: email
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: sortBy
          schema: { type: string }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags: [users]
      summary: 创建用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      tags: [users]
      summary: 获取用户详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [users]
      summary: 更新用户
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [users]
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/roles:
    post:
      tags: [users]
      summary: 为用户分配角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
          description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_ids]
              properties:
                role_ids:
                  type: array
                  items: { type: integer }
                  description: 要分配给用户的角色ID列表
      responses:
        '200':
          description: 角色分配成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 错误的请求 (例如，无效的用户ID格式，无效的角色ID，或格式错误的请求)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [users]
      summary: 从用户移除角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
          description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_ids]
              properties:
                role_ids:
                  type: array
                  items: { type: integer }
                  description: 要从用户移除的角色ID列表
      responses:
        '200':
          description: 角色移除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 错误的请求 (例如，无效的用户ID格式，无效的角色ID，或格式错误的请求)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 角色管理
  /roles:
    get:
      tags: [roles]
      summary: 获取角色列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
    post:
      tags: [roles]
      summary: 创建角色
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /roles/{roleId}:
    get:
      tags: [roles]
      summary: 获取角色详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailsResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [roles]
      summary: 更新角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [roles]
      summary: 删除角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体
        '400':
          description: 角色有关联用户，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /roles/{roleId}/permissions:
    get:
      tags: [roles]
      summary: 获取角色的权限列表
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
    post:
      tags: [roles]
      summary: 分配权限到角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignPermissionsRequest'
      responses:
        '200':
          description: 分配成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /permissions:
    get:
      tags: [permissions]
      summary: 获取权限列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 50 }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'

  # 职责管理
  /responsibilities:
    get:
      tags: [responsibilities]
      summary: 获取职责列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsibilityListResponse'
    post:
      tags: [responsibilities]
      summary: 创建职责
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsibilityCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsibilityResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /responsibilities/{respId}:
    get:
      tags: [responsibilities]
      summary: 获取职责详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: respId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsibilityDetailsResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [responsibilities]
      summary: 更新职责
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: respId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsibilityUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsibilityResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [responsibilities]
      summary: 删除职责
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: respId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体
        '400':
          description: 有关联用户，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 环境管理
  /environments:
    get:
      tags: [environments]
      summary: 获取环境列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: code
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
    post:
      tags: [environments]
      summary: 创建环境
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
  /environments/{envId}:
    get:
      tags: [environments]
      summary: 获取环境详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentDetailsResponse'
    put:
      tags: [environments]
      summary: 更新环境
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
    delete:
      tags: [environments]
      summary: 删除环境
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体

  # 资产管理
  /assets:
    get:
      tags: [assets]
      summary: 获取资产列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: ipAddress
          schema: { type: string }
        - in: query
          name: hostname
          schema: { type: string }
        - in: query
          name: os
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
    post:
      tags: [assets]
      summary: 创建资产
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
...
  /audit-logs/{auditLogId}:
    get:
      tags: [audit-logs]
      summary: 获取审计日志详情
      description: 获取指定ID的审计日志详细信息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: auditLogId
          required: true
          schema: { type: integer }
          description: 审计日志ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 服务实例管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserCreateRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        department: { type: string }
        roles:
          type: array
          items: { type: integer }
...
    UserListResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/User' }
            total: { type: integer }
            page: { type: integer }
            pageSize: { type: integer }
...
    UserMeResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          $ref: '#/components/schemas/AuthMeData'
    # 角色、权限、职责、环境、资产、服务、业务、Bug、审计日志等 schemas 自动补全，详见前述设计文档结构
    # ...
...
    SuccessResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data: { type: [string, "null"] }
    ErrorResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: [string, "null"] }
    AuthLoginResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: '#/components/schemas/User'
...